[Plugins] [UDocker] fix: minimise potential race conditions in sharedContainer mode by assigning distinct containers to concurrent threads
